+++ engine/source/platformOSX/platformOSX.mm
@@ -247,7 +247,7 @@ static osxPlatState * tempSharedPlatState = nil;
 
 //-----------------------------------------------------------------------------
 
-- (unsigned)retainCount
+- (NSUInteger)retainCount
 {
     // Denotes an object that cannot be released
     return UINT_MAX;
